Refactor: Move getDriversByIds to userServiceClient integration
